requirements.txt
# --- Core LLM & Orchestration ---
langchain>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.0

# --- Embeddings & Vector Store ---
langchain-huggingface>=0.1.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0

# --- PDF Processing ---
langchain-community>=0.3.0
pypdf>=4.0.0

# --- Animation ---
manim>=0.18.0

# --- YouTube ---
google-api-python-client>=2.100.0

# --- UI ---
streamlit>=1.38.0

# --- Utilities ---
python-dotenv>=1.0.0
